+++ src/player/main/sp/it/pl/main/AppDrag.kt
@@ -165,15 +165,15 @@ fun Clipboard.setSongsAndFiles(items: List<Song>) {
  *
  * @return true iff contains at least 1 audio file or audio url or (any) directory
  */
-fun Dragboard.hasAudio(): Boolean = Df.SONGS in this || (hasUrl() && url.isAudio()) || (hasFiles() && files.any { it.isAudio() || it.isDirectory })
+fun Dragboard.hasAudio(): Boolean = Df.SONGS in this || (hasUrl() && url.isAudioOrVideo()) || (hasFiles() && files.any { it.isAudioOrVideo() || it.isDirectory })
 
 /** @return list of songs as specified in [Dragboard.hasAudio] */
 fun Dragboard.getAudio(): List<Song> = when {
    Df.SONGS in this -> this[Df.SONGS]
-   hasFiles() -> findAudio(files).map { SimpleSong(it) }.toList()
+   hasFiles() -> findAudioOrVideo(files).map { SimpleSong(it) }.toList()
    hasUrl() -> {
       when {
-         url.isAudio() -> listOf(SimpleSong(uri(url)))
+         url.isAudioOrVideo() -> listOf(SimpleSong(uri(url)))
          else -> listOf()
       }
    }